requirements.txt
numpy>=1.23.5,<=1.26.4
scipy>=1.11,<1.17
tensorflow==2.15.1
pyyaml
scikit-learn>=1.5.2,<=1.7.0
pandas>=2.0
cerberus
matplotlib>=3.9.0
optuna>=4.0
chardet>=5.0